+++ R/sim_program_msms.R
@@ -7,7 +7,7 @@
 # Licence:  GPLv3 or later
 # --------------------------------------------------------------
 
-callMsms <- function(jar.path, ms.args, msms.args) {
+callMsms <- function(jar.path, ms.args, msms.args, subgroup) {
   out.file = getTempFile("msms")
   seed <- generateSeeds(1)
 
@@ -78,7 +78,7 @@ generateMsmsOptionsCommand <- function(dm) {
   cmd
 }
 
-generateMsmsOptions <- function(dm, parameters) {
+generateMsmsOptions <- function(dm, parameters, subgroup) {
   msms.tmp <- new.env()
 
   par.names <- dm.getParameters(dm)
@@ -109,15 +109,20 @@ msmsSimFunc <- function(dm, parameters) {
     stop("Wrong number of parameters!")
 
   # Generate Options
-  ms.options <- paste(sum(dm.getSampleSize(dm)), dm.getLociNumber(dm), 
-                      paste(generateMsOptions(dm, parameters), collapse=" "))
-  msms.options <- paste(generateMsmsOptions(dm, parameters), collapse= " ") 
-
-  # Simulate
-  out.file <- callMsms(getJaathaVariable('msms.jar'), ms.options, msms.options)
-
-  # Generate & return summary statistic
-  generateSumStats(out.file, 0, parameters, dm)
+  subgroup_sizes <- sampleSubgroupSizes(dm)
+  msms.files <- sapply(1:dm.getSubgroupNumber(dm), function(subgroup) {
+    if (subgroup_sizes[subgroup] == 0) return("")
+    ms.options <- paste(sum(dm.getSampleSize(dm)),
+                        subgroup_sizes[subgroup],
+                        paste(generateMsOptions(dm, parameters, subgroup), 
+                              collapse=" "))
+    msms.options <- paste(generateMsmsOptions(dm, parameters, subgroup), 
+                          collapse= " ")
+    callMsms(getJaathaVariable('msms.jar'), ms.options, msms.options)
+  })
+  
+  # Parse the simulation output
+  generateSumStats(msms.files, 0, parameters, dm)
 }
 
 finalizeMsms <- function(dm) {