+++ R/sum_stat_generate.R
@@ -1,41 +1,41 @@
-generateSumStats <- function(file, program, parameters, dm) {
+generateSumStats <- function(files, program, parameters, dm, seg_sites) {
   model_stats <- dm.getSummaryStatistics(dm)
   
-  # Detemine if we need to generate seg.sites
-  generate_seg_sites <- 'seg.sites' %in% model_stats
-  if (any(c('fpc', 'pmc') %in% model_stats)) {
-    generate_seg_sites <- TRUE
+  calc_seg_sites <- any(c('seg.sites', 'jsfs', 'pmc', 'fpc') %in% model_stats)
+  if (missing(seg_sites) & calc_seg_sites) {
+    seg_sites <- parseOutput(files, dm.getSampleSize(dm), 
+                             dm.getLociNumber(dm), program, 
+                             dm.getLociTrioOptions(dm))
+  }
+
+  # Add the parameters of the simulation
+  sum_stats <- list(pars=parameters)
+
+  # Add seg_sites
+  if ('seg.sites' %in% model_stats) {
+    sum_stats[['seg.sites']] <- seg_sites
   }
   
-  # Parse the simulation output
-  sum_stats <- parseOutput(file, dm.getSampleSize(dm), dm.getLociNumber(dm), 
-                           program, 'jsfs' %in% model_stats, generate_seg_sites,
-                           dm.getLociTrioOptions(dm))
-  
-  # Add the parameters of the simulation
-  sum_stats[['pars']] <- parameters
+  # Add JSFS
+  if ('jsfs' %in% model_stats) {
+    sum_stats[['jsfs']] <- calcJsfs(seg_sites, dm.getSampleSize(dm))
+  }
   
   # Add pmc statistic if needed
   if ('pmc' %in% model_stats) {
-    stopifnot(!is.null(sum_stats$seg.sites))
-    sum_stats[['pmc']] <- createPolymClasses(sum_stats$seg.sites, dm)
+    sum_stats[['pmc']] <- createPolymClasses(seg_sites, dm)
   }
   
   # Add fpc statistic if needed
   if ('fpc' %in% model_stats) {
-    stopifnot(!is.null(sum_stats$seg.sites))
-    sum_stats[['fpc']] <- generateFpcStat(sum_stats$seg.sites, dm)
+    sum_stats[['fpc']] <- generateFpcStat(seg_sites, dm)
   }
   
-  # Remove seg.sites if it was just there to generate pmc or fpc.
-  if (!'seg.sites' %in% model_stats) sum_stats[['seg.sites']] <- NULL
-  sum_stats  
-  
-  # Add the simulation file if needed, or delete it otherwise
-  if ("file" %in% model_stats) {
-    sum_stats[['file']] <- file
+  # Add files if needed
+  if ('file' %in% model_stats) {
+    sum_stats[['file']] <- files
   } else {
-    unlink(file)
+    unlink(files)
   }
   
   sum_stats