+++ src/sum_stat_jsfs.cpp
@@ -1,34 +1,30 @@
 #include <Rcpp.h>
 using namespace Rcpp;
 
-// Fast function to add a locus' seg.sites to a JSFS
-void addToJsfs(const NumericMatrix seg_sites,
-               const NumericVector sample_size,
-               NumericMatrix &jsfs) {
+//' Calculate the JSFS from a list of segregating sites statistics
+//' 
+//' @param seg_sites List of segregating sites
+//' @param sample_size A numberic vector of size 2, giving the sample sizes
+//'        of the two population for which the JSFS is calulated
+//' @return The Joint Site Frequency Spectrum, as a matrix.
+// [[Rcpp::export]]
+NumericMatrix calcJsfs(const List seg_sites, const NumericVector sample_size) {
   
-  int idx1;
-  int idx2;
+  NumericMatrix jsfs(sample_size[0]+1, sample_size[1]+1);
+  size_t idx1, idx2;
   
-  for (int j = 0; j < seg_sites.ncol(); ++j) {
-    idx1 = 0;
-    idx2 = 0;
+  for (int locus = 0; locus < seg_sites.size(); ++locus) {
+    NumericMatrix ss = as<NumericMatrix>(seg_sites[locus]);
+    for (int j = 0; j < ss.ncol(); ++j) {
+      idx1 = 0;
+      idx2 = 0;
     
-    for (int i = 0; i < sample_size[0]; ++i) idx1 += seg_sites(i,j); 
-    for (int i = sample_size[0]; i < seg_sites.nrow(); ++i) idx2 += seg_sites(i,j); 
+      for (int i = 0; i < sample_size[0]; ++i) idx1 += ss(i,j); 
+      for (int i = sample_size[0]; i < ss.nrow(); ++i) idx2 += ss(i,j); 
     
     ++jsfs(idx1, idx2);
+    }
   }
-}
-
-// Exportable wrapper function for unit testing
-// [[Rcpp::export]]
-NumericMatrix addSegSitesToJsfs(const NumericMatrix seg_sites, 
-                                const NumericVector sample_size,
-                                NumericMatrix jsfs) {
   
-  NumericMatrix jsfs_copy = clone(jsfs);
-  addToJsfs(seg_sites, sample_size, jsfs_copy);
-  
-  
-  return jsfs_copy;
+  return jsfs;
 }