+++ tests/testthat/test-u-cpp-parse-output.R
@@ -14,23 +14,62 @@ test_that("parseMsPositions works", {
   expect_error(parseMsPositions("segsites: 0"))
 })
 
-test_that("test.parseOutput.ms works", {
+test_that("parseOutput works for ms", {
   set.seed(25)
   dm.tt@sum.stats <- data.frame()
   dm.tt <- dm.addSummaryStatistic(dm.tt, "file")
-  dm.tt <- dm.addSummaryStatistic(dm.tt, "seg.sites")
   ss <- dm.getSampleSize(dm.tt)
   ln <- dm.getLociNumber(dm.tt)
-  breaks <- c(0.25, 0.5, 0.75)
+
   ms.file <- dm.simSumStats(dm.tt, c(1, 5))$file
   expect_error(parseOutput("bulb.txt", ss, ln))
-  sum.stats <- parseOutput(ms.file, ss, ln, 0, FALSE, TRUE, FALSE)
-  expect_equal(length(sum.stats), 1)
-  expect_true(is.list(sum.stats$seg.sites))
-  expect_equal(length(sum.stats$seg.sites), dm.getLociNumber(dm.tt))
-  sum.stats <- parseOutput(ms.file, ss, ln, 0, TRUE, TRUE, FALSE)
-  expect_equal(length(sum.stats), 2)
-  expect_true(is.list(sum.stats$seg.sites))
-  expect_true(is.matrix(sum.stats$jsfs))
-  expect_true(sum(sum.stats$jsfs) > 0)
+  
+  seg_sites <- parseOutput(ms.file, ss, ln, 0)
+  expect_true(is.list(seg_sites))
+  expect_equal(length(seg_sites), dm.getLociNumber(dm.tt))
+  for (seg_site in seg_sites) {
+    expect_true(is.matrix(seg_site))
+    expect_equal(nrow(seg_site), sum(ss))
+    expect_true(all(seg_site %in% c(0, 1)))
+  }
+  
+  ms.file <- c(ms.file, dm.simSumStats(dm.tt, c(1, 5))$file)
+  seg_sites <- parseOutput(ms.file, ss, 2*ln, 0)
+  expect_true(is.list(seg_sites))
+  expect_equal(length(seg_sites), 2*dm.getLociNumber(dm.tt))
+  for (seg_site in seg_sites) {
+    expect_true(is.matrix(seg_site))
+    expect_equal(nrow(seg_site), sum(ss))
+    expect_true(all(seg_site %in% c(0, 1)))
+  }
+  
+  unlink(ms.file)
+})
+
+test_that("parseOutput work for seq-gen", {
+  if (!test_seqgen) return()
+  invisible(files <- dm.simSumStats(dm.hky, c(1, 5))$file)
+  
+  seg_sites <- parseOutput(files[['seqgen']], dm.getSampleSize(dm.hky), 
+                           dm.getLociNumber(dm.hky), 1)
+  expect_true(is.list(seg_sites))
+  expect_equal(length(seg_sites), dm.getLociNumber(dm.hky))
+  for (seg_site in seg_sites) {
+    expect_true(is.matrix(seg_site))
+    expect_equal(nrow(seg_site), sum(dm.getSampleSize(dm.hky)))
+    expect_true(all(seg_site %in% c(0, 1)))
+  }
+  
+  seg_sites <- parseOutput(c(files[['seqgen']], files[['seqgen']]), 
+                           dm.getSampleSize(dm.hky), 
+                           2*dm.getLociNumber(dm.hky), 1)
+  expect_true(is.list(seg_sites))
+  expect_equal(length(seg_sites), 2*dm.getLociNumber(dm.hky))
+  for (seg_site in seg_sites) {
+    expect_true(is.matrix(seg_site))
+    expect_equal(nrow(seg_site), sum(dm.getSampleSize(dm.hky)))
+    expect_true(all(seg_site %in% c(0, 1)))
+  }
+  
+  unlink(files)
 })